+++ tests/serde_tests/mod.rs
@@ -1,0 +1,2 @@
+mod doc_serde;
+mod internals_erased;